+++ demo/index.ts
@@ -1,10 +1,10 @@
-import { InDiv, Component, Utils, NvModule, OnInit, DoCheck, BeforeMount, AfterMount, ReceiveInputs, SetState, OnDestory, setState, ElementRef, Watch, HasRender, Input } from '@indiv/core';
+import { InDiv, Component, Utils, NvModule, OnInit, DoCheck, BeforeMount, AfterMount, ReceiveInputs, SetState, OnDestory, setState, ElementRef, HasRender, Input, ViewChild, ViewChildren } from '@indiv/core';
 import { RouteChange, NvLocation, RouteModule } from '@indiv/router';
 import { PlatformBrowser } from '@indiv/platform-browser';
 import { HttpClient, HttpClientResponse } from '@indiv/common';
 import { Observable } from 'rxjs';
 
-import { SharedModule } from './share.module';
+import { SharedModule, TestDirective } from './share.module';
 import { HeroSearchService, HeroSearchService1, HeroSearchService2 } from './service';
 import { PrivateService } from './private.service';
 
@@ -229,7 +229,7 @@ class R2 implements OnInit, BeforeMount, AfterMount, DoCheck, RouteChange, OnDes
       <p nv-on:click="click()">测试repeat组件: {{manName}}</p>
     </div>`),
 })
-class TestComponent implements OnDestory, ReceiveInputs {
+class TestComponent implements OnDestory, ReceiveInputs, AfterMount, HasRender {
   public state: any;
   @Input() public manName: any;
 
@@ -249,6 +249,13 @@ class TestComponent implements OnDestory, ReceiveInputs {
     this.manName = 'fuck!';
   }
 
+  public nvHasRender() {
+    console.log('TestComponent HasRender');
+  }
+
+  public nvAfterMount() {
+    console.log('TestComponent AfterMount');
+  }
   public nvOnDestory() {
     console.log('TestComponent OnDestory');
   }
@@ -263,7 +270,7 @@ class TestComponent implements OnDestory, ReceiveInputs {
   template: (`
   <div class="fucck" nv-class="test.a" nv-id="'cc'">
     <input nv-model="test.a" nv-on:click="show(test)" />
-    <p test-directive="{'123'}" nv-id="232" nv-if="countState(a)" nv-on:click="changeInput()">{{a}}</p>
+    <p test-directive="{test.a}" nv-id="232" nv-if="countState(a)" nv-on:click="changeInput()">{{a}}</p>
     <test-component nv-repeat="man in testArray" nv-key="man.name" manName="{countState(man.name)}" nv-if="a"></test-component>
     <p nv-on:click="go()">container: {{countState(color)}}</p>
     <input type="number" nv-model="a" />
@@ -348,6 +355,11 @@ class Container implements OnInit, AfterMount, DoCheck, HasRender, RouteChange {
   public setState: SetState;
   public http$: Observable<HttpClientResponse>;
 
+  @ViewChild('test-component') private testComponent: TestComponent;
+  @ViewChild('router-render') private routerRenderElementRef: ElementRef;
+  @ViewChildren('test-directive') private testDirectiveString: TestDirective[];
+  @ViewChildren(TestDirective) private testDirective: TestDirective[];
+
   constructor(
     private hss: HeroSearchService,
     private value: ValueType,
@@ -366,9 +378,9 @@ class Container implements OnInit, AfterMount, DoCheck, HasRender, RouteChange {
       };
     });
     this.http$ = this.httpClient.get('/success');
-    this.http$.subscribe({
-      next: this.httpHandler,
-    });
+    // this.http$.subscribe({
+    //   next: this.httpHandler,
+    // });
     this.hss.test();
     console.log('value', this.value);
     setTimeout(() => {
@@ -388,13 +400,12 @@ class Container implements OnInit, AfterMount, DoCheck, HasRender, RouteChange {
     console.log('nvOnInit Container', this.location.get());
   }
 
-  public nvAfterMount() {
-    console.log('nvAfterMount Container');
-    // document.getElementById('1').className = '3333';
+  public nvHasRender() {
+    console.log('nvHasRender Container', 33333333, this.testComponent, this.testDirective, this.routerRenderElementRef, this.testDirectiveString);
   }
 
-  public nvHasRender() {
-    console.log('nvHasRender Container');
+  public nvAfterMount() {
+    console.log('nvAfterMount Container', 222222, this.testComponent, this.testDirective, this.routerRenderElementRef, this.testDirectiveString);
   }
 
   public go() {