+++ packages/core/component/utils.ts
@@ -82,7 +82,7 @@ function resolveDirective(componentInstance: IComponent): void {
     } else {
       let pureProps = null;
       if (/^nv-repeat=.*/.test(matchProps)) {
-        const _value = pureMatchProps.split('in')[1];
+        const _value = pureMatchProps.split(' in ')[1];
         if (!_value) throw new Error(`directive nv-repeat 's expression ${pureMatchProps} is wrong!`);
         const value = _value.replace(/\s*/g, '');
         pureProps = value.split('.')[0];