+++ packages/platform-server/plugins/to-string.ts
@@ -1,5 +1,5 @@
 import { InDiv, INvModule } from '@indiv/core';
-import { TRouter, nvRouteStatus } from '@indiv/router';
+import { TRouter, NvLocation } from '@indiv/router';
 import { _document } from '../renderer';
 import { PlatformServer } from './platform-server';
 import { buildPath, generalDistributeRoutes } from '../router';
@@ -11,24 +11,29 @@ import { buildPath, generalDistributeRoutes } from '../router';
  *
  * @export
  * @param {Function} rootModule
- * @param {string} [url]
- * @param {{ routes?: TRouter[], nvRootPath?: string }} [routeConfig]
+ * @param {{
+ *   path?: string,
+ *   query?: any,
+ *   routes?: TRouter[],
+ *   nvRootPath?: string,
+ * }} [routeConfig]
  * @returns {Promise<string>}
  */
-export async function renderToString(rootModule: Function, url?: string, routeConfig?: { routes?: TRouter[], nvRootPath?: string }): Promise<string> {
+export async function renderToString(rootModule: Function, routeConfig?: {
+  path?: string,
+  query?: any,
+  routes?: TRouter[],
+  nvRootPath?: string,
+}): Promise<string> {
   if (_document.getElementById('root')) _document.getElementById('root').innerHTML = '';
   const inDiv = new InDiv();
   inDiv.bootstrapModule(rootModule);
   inDiv.use(PlatformServer);
   await inDiv.init();
-  if (url && routeConfig) {
-    nvRouteStatus.nvRootPath = routeConfig.nvRootPath;
-    nvRouteStatus.nvRouteObject = {
-      path: url,
-      query: null,
-      data: null,
-    };
-    const renderRouteList = buildPath(url);
+  if (routeConfig) {
+    const nvLocatiton = new NvLocation(inDiv);
+    nvLocatiton.set(routeConfig.path, routeConfig.query);
+    const renderRouteList = buildPath(routeConfig.path);
     const routesList: TRouter[] = [];
     const loadModuleMap: Map<string, INvModule> = new Map();
     await generalDistributeRoutes(routeConfig.routes, routesList, renderRouteList, inDiv, loadModuleMap);