+++ packages/router/route/router-module.ts
@@ -1,4 +1,4 @@
-import { IComponent, IDirective, INvModule, ComponentList, DirectiveList, factoryModule, NvModule, InDiv, utils, Vnode } from '@indiv/core';
+import { IComponent, INvModule, ComponentList, DirectiveList, factoryModule, NvModule, InDiv, Vnode, utils } from '@indiv/core';
 
 import { nvRouteStatus, NvLocation } from './location';
 import { RouterTo, RouterFrom } from './directives';
@@ -74,7 +74,7 @@ export class RouteModule {
     this.indivInstance.setRouteDOMKey('router-render');
 
     // if isn't browser, will auto watch nvRouteStatus.nvRouteObject
-    if (!utils.isBrowser()) return;
+    if (!utils.hasWindowAndDocument()) return;
 
     // if is browser, will watch nvRouteStatus.nvRouteObject by 'load' event of window
     window.addEventListener('load', () => this.refresh(), false);
@@ -428,11 +428,11 @@ export class RouteModule {
    * emit nvRouteChange and nvOnDestory for Components with recursion
    * 
    * @private
-   * @param {ComponentList<IComponent>[]} componentList
+   * @param {ComponentList[]} componentList
    * @param {string} event
    * @memberof Router
    */
-  private emitComponentEvent(componentList: ComponentList<IComponent>[], event: string): void {
+  private emitComponentEvent(componentList: ComponentList[], event: string): void {
     if (event === 'nvRouteChange') {
       componentList.forEach(component => {
         if (component.instanceScope.nvRouteChange) component.instanceScope.nvRouteChange(this.lastRoute, this.currentUrl);
@@ -453,11 +453,11 @@ export class RouteModule {
    * emit nvRouteChange and nvOnDestory for Directives with recursion
    *
    * @private
-   * @param {DirectiveList<IDirective>[]} directiveList
+   * @param {DirectiveList[]} directiveList
    * @param {string} event
    * @memberof RouteModule
    */
-  private emitDirectiveEvent(directiveList: DirectiveList<IDirective>[], event: string): void {
+  private emitDirectiveEvent(directiveList: DirectiveList[], event: string): void {
     if (event === 'nvRouteChange') {
       directiveList.forEach(directive => {
         if (directive.instanceScope.nvRouteChange) directive.instanceScope.nvRouteChange(this.lastRoute, this.currentUrl);