+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### Features
 
+* **package:** add @ViewChild and @ViewChildren in @indiv/core
+* **package:** optimize @Input
+
+### Fix
+
+* **package:** fix compile `nv-repeat` in @indiv/core
+* **package:** fix nvLocation parmas in @indiv/platform-server
+
+
+<a name="0.0.2-alpha.0"></a>
+# [0.0.2-alpha.0](https://github.com/DimaLiLongJi/InDiv/releases/tag/v0.0.2-alpha.0) (2019-01-11)
+
+
+### Features
+
 * **package:** optimize render in @indiv/core
 * **package:** add package of server side render: @indiv/platform-server